feat: refactor CarouselCards component; remove HoverSparkles and improve layout styles; add Footer component
