+++ components/Animation/money.js
