+++ pages/index.js
@@ -1,7 +1,6 @@
 import Head from 'next/head'
 import Carousel from '../components/index/Carousel'
 import CreateCard from '../components/index/cards/CreateCard'
-import Sparkles from '../components/sparkles'
 import cards from '../components/index/cards/cards'
 
 export default function Home() {
@@ -11,7 +10,6 @@ export default function Home() {
         <title>HackClub Butwal</title>
         <meta name="description" content="HackClub Butwal is launching soon. Stay tuned!" />
       </Head>
-      <Sparkles />
       <Carousel cards={cards} />
       <CreateCard />
     </div>
